+++ crates/vfs/Cargo.toml
@@ -4,5 +4,7 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
+utils.workspace = true
+
 thunderdome.workspace = true
 walkdir = "2"
